requirements.txt
langchain==0.3.21
langchain-community==0.3.20
langchain-openai==0.3.9 
python-dotenv==1.0.1
streamlit==1.43.2
langgraph==0.3.18
psycopg2-binary==2.9.9